requirements.txt
audio_recorder_streamlit==0.0.8
ipython==8.12.2
langchain==0.0.271
librosa==0.10.0.post2
numpy==1.24.3
openai==0.27.8
pandas==2.0.3
Pillow==9.4.0
Pillow==10.0.0
plotly==5.15.0
Requests==2.31.0
scikit_learn==1.3.0
soundfile==0.12.1
speechbrain==0.5.15
streamlit==1.24.1
streamlit_chat==0.1.1
streamlit_extras==0.3.0
torch==2.0.1
torchaudio==2.0.2
web3==6.8.0
